Add admin privilege checks and restructure code into Providers namespace.
