+++ Flow.Launcher.Plugin.FlowGet/Providers/UiExecutor.cs
@@ -2,7 +2,7 @@
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace Flow.Launcher.Plugin.FlowGet;
+namespace Flow.Launcher.Plugin.FlowGet.Providers;
 
 internal class UiExecutor(IPublicAPI? api, string pluginName)
 {
